DRY test helpers
