+++ test/test_helper.exs
@@ -5,3 +5,5 @@ Mix.Task.run "ecto.migrate", ["--quiet"]
 Ecto.Adapters.SQL.begin_test_transaction(Gatekeeper.Repo)
 
 ExUnit.configure(exclude: [skip: true])
+
+Code.load_file "test/factory.exs"
